python-internal/pyproject.toml
[project]
name = "python-internal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "flask>=3.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
